backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
pinecone-client==3.2.2
groq==0.4.2
httpx==0.27.0
python-dotenv==1.0.1
